Update unlink, seek

+++ lib/libfs.h
@@ -23,7 +23,7 @@ int libfs_unlink(char *const name);
 fd_type libfs_open(char *const name, const int flags);
 int libfs_read(const fd_type fd, char *buf, const unsigned int size);
 int libfs_write(const fd_type fd, const char *buf, const unsigned int size);
-int libfs_seek(const fd_type fd, const long int offset);
+int libfs_seek(const fd_type fd, const unsigned offset);
 int libfs_close(const fd_type fd);
 int libfs_stat(const char *path, struct stat_t *buf);
 int libfs_link(const char *oldpath, const char *newpath);
